refactor(ui): split TryItOutForm into linear/ + variant switcher
Move the existing onboarding form into try-it-out-form/linear/ and
extract shared sx helpers into try-it-out-form/styles.ts. The top-level
TryItOutForm becomes a thin switcher (currently always renders the
linear variant) to make room for the upcoming wizard variant.

No behavior change.
